docker::stack unless and onlyif were doing "docker stack deploy ls"
"docker stack deploy ls" is not a command so it would fail every time, returning exit code of 1

This means the docker stack deploy that creates the stack will always run. This may have a positive side effect b/c docker stack deploy will create or update the stack, so maybe if the compose file has changed it will detect that and recreate it? It doesn't fail even if the stack is already created. Maybe the unless should just be removed so it will create/update every puppet run? Or maybe it should be refreshonly and another exec could notify it if the stack doesn't exist? That way the refreshonly exec could be notified by whoever is managing the docker-compose.yaml file?

I assume the onlyif on exec that does "docker stack rm ${stackname}" will never run because the onlyif (doing "docker stack deploy ls") will always return 1 due to invalid command.