(MODULES-9036) Fix Windows acceptance tests
While fixing the tests I uncovered that the docker_stack provider does
not check for the existence of stacks properly on Windows. This appears
to be due to it looking for containers tagged with latest, as not having
any tag. As a result, the stack is created on every run and also cannot
be destroyed. The changes here help rectify this however there is
probably a better solution needed.