(MODULES-9193) Revert part of MODULES-9177
The changes introduced in MODULES-9177 modified the behaviour of docker
installs using older versions. The first part of the modified regex ensured
that the correct repository was used. As such I have reverted it but
left the rest of the changes to ensure that non 17.y.z versions use the
correct repo.