(GH-1097) Add mint support
This allows the docker module to be installed on Mint systems, as though
they were ubuntu systems, without having to change Mint system facts.
Specifically it changes the OS names when installing pre-required
packages so that the default for all Debian family OSes is the same as
the Debian required packages, and so that Mint and Ubuntu install the
same packages.