allow docker::networks::networks param to be undef
I would like to include the docker::networks class into the role
(manifest) of nodes which are also including the docker module but not
necessarily always defining docker network(s). This would allow docker
networks to optionally be defined, when needed, via hiera. Currently,
this type of arrangement would require constructing a wrapper class as
the docker::networks::networks parameter is mandatory.