Attempt to escape things in on OS dependant way
Ruby Shellwords#shellescape "Escapes a string so that it can be safely
used in a Bourne shell command line".  This is fine for a UNIX like
operating system, but windows does not complies with this.

Attempt to fix CI by manually "escaping" strings on windows (i.e. quote
them and escape quotes with a backslash), and relying on shellescape for
other operating systems.